Small project file changes.
git-svn-id: http://dminfo:8686/svn/repository/models/trunk@484 cdc4813b-4270-ec4f-936e-925f93a782c2

+++ dsm2/src/oprule_interface/dsm2_named_value_factories.cpp
@@ -103,7 +103,7 @@ void convert_reservoir_identifiers(
    }
 
    string name=iter->second;
-   resno=reservoir_index(name.c_str(),name.length());
+   resno=(int)reservoir_index(name.c_str(),name.length());
    iter = argmap.find("connect");
    if (iter == argmap.end() ){
 	   if (conn_requested){
